+++ utility.c
@@ -6,13 +6,13 @@
 /*   By: releanor <releanor@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/02 12:55:23 by releanor          #+#    #+#             */
-/*   Updated: 2020/03/02 22:40:23 by releanor         ###   ########.fr       */
+/*   Updated: 2020/03/06 11:11:40 by releanor         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int		num_len(int num, int base)
+int		num_len(intmax_t num, int base)
 {
 	int		num_len;
 
@@ -26,9 +26,9 @@ int		num_len(int num, int base)
 	return (num_len);
 }
 
-int		unsigned_num_len(unsigned int num, int base)
+int		unsigned_num_len(uintmax_t num, int base)
 {
-	int		num_len;
+	uintmax_t		num_len;
 
 	num_len = 0;
 	if (num == 0)
